update: move POST coupons to admin-specific methods
