+++ lib/api/discount-coupons.ts
@@ -11,16 +11,6 @@ async function getCouponByCouponNumber(
   return response;
 }
 
-async function createCoupons(
-  couponsData: DiscountCoupon[],
-  axiosWithAuth: AxiosInstance
-): Promise<ApiResponse<DiscountCoupon[]>> {
-  const response = (await axiosWithAuth.post(`/discount-coupons`, {
-    coupons: couponsData,
-  })) as ApiResponse<DiscountCoupon[]>;
-  return response;
-}
-
 async function updateCoupon(
   couponData: DiscountCoupon,
   axiosWithAuth: AxiosInstance
@@ -33,6 +23,5 @@ async function updateCoupon(
 
 export const discountCoupons = {
   getCouponByCouponNumber,
-  createCoupons,
   updateCoupon,
 };
